[BACKEND] Refactore RemoveLayoutConversion pass
Significant changes to the pass logic. Move away from greedy rewrites
and use more global analysis instead. The pass is now bocken down into
2 main phases. First forward propagation of layout starting from ops
that we don't want to change. Propagate to all the nodes.
If there is a single layout needed for the op then we can rewrite the op,
if there are multiple layout required based on dependency we need a
tie break.
The second phase is backward propgation that gets a backward slice of operations
starting from the convert and if all the operations in the slice can be
rematerialized rewrite the slice. This backward phase now supports going through
loop arguments.

This will allow more complex logic in the future to add a cost model to
decide which convert to leave and which to fold