[BACKEND] Promote lhs to tmem more aggressively
Extend the logic to promote lhs to tmem, until now we were doing it only
if the layout allowed copying to tmem but often we can change the layout
and the convert_layout will propagate up.
Change the heuritstic to promote to tmem as long as the value doesn't
come from a global load, this matches the logic on Hopper.